slot controller, service , fhir validator modification
